# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Psychology Clinic API",
    "description": "REST API for managing fictitious psychology clinic patients and sessions. Created for CSE 341 Project 2.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/github/callback": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "GitHub OAuth callback",
        "description": "Callback used by GitHub after successful authentication.",
        "responses": {
          "302": {
            "description": "Redirects to the authenticated user profile"
          },
          "401": {
            "description": "GitHub authentication failed"
          }
        }
      }
    },
    "/auth/profile": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get authenticated user",
        "description": "Returns the currently authenticated GitHub user.",
        "responses": {
          "200": {
            "description": "Authentication successful"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
...
  }
}